Use only exported types
